cloud-node/requirements.txt
absl-py==0.8.1
appnope==0.1.0
astor==0.8.0
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
backcall==0.1.0
blessed==1.15.0
boto3==1.9.130
botocore==1.12.136
cached-property==1.5.1
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.3.9
constantly==15.1.0
decorator==4.4.0
dnspython==1.16.0
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
eventlet==0.24.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SocketIO==3.3.2
Flask-Sockets==0.2.1
Flask-uWSGI-WebSocket==0.6.1
future==0.16.0
gast==0.2.2
gevent==1.4.0
gevent-websocket==0.10.1
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.24.3
h5py==2.10.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mock==2.0.0
monotonic==1.5
numpy==1.17.3
opt-einsum==3.1.0
parso==0.4.0
pathspec==0.5.9
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==1.0.14
protobuf==3.10.0
ptyprocess==0.6.0
pyasn1==0.4.5
Pygments==2.4.2
PyHamcrest==1.9.0
PyInquirer==1.0.3
python-dateutil==2.8.0
python-engineio==3.5.0
python-socketio==3.1.2
PyYAML==3.13
pyzmq==18.0.1
regex==2019.11.1
requests==2.20.1
rsa==3.4.2
s3transfer==0.2.0
scipy==1.2.1
semantic-version==2.5.0
SimpleWebSocketServer==0.1.1
six==1.12.0
tb-nightly==1.14.0a20190319
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-estimator-2.0-preview==1.14.0.dev2019040400
tensorflow-hub==0.5.0
tensorflowjs==1.3.1.1
termcolor==1.1.0
texttable==0.9.1
tornado==6.0.2
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
uWSGI==2.0.18
wcwidth==0.1.7
websocket-client==0.56.0
websocket-server==0.4
websockets==7.0
Werkzeug==0.16.0
wrapt==1.11.2
zope.interface==4.6.0
